Dependeny-Updates und Refactorings
