+++ src/features/landingpage/landing-page.tsx
@@ -1,0 +1,7 @@
+export const LandingPage = () => {
+    return (
+        <>
+            Landing page
+        </>
+    )
+}
